Implemented vue-router, set up 404 path, redirects to home path, converted Card.vue container elements to RouterLinks with dynamic quiz id trailing path.
