+++ .vscode/settings.json
@@ -1,0 +1,7 @@
+{
+    "workbench.colorCustomizations": {
+        "activityBar.background": "#362D0C",
+        "titleBar.activeBackground": "#4B3F11",
+        "titleBar.activeForeground": "#FCF9F1"
+    }
+}
