+++ QuizApp/src/router/index.js
@@ -1,0 +1,32 @@
+import { createRouter, createWebHistory } from "vue-router";
+import HomeView from "../views/HomeView.vue";
+import PageNotFoundView from "../views/404View.vue";
+import QuizView from "../views/QuizView.vue";
+
+const router = createRouter({
+    history: createWebHistory(import.meta.env.BASE_URL),
+    routes: [
+        {
+            path: "/",
+            name: "home",
+            component: HomeView
+        },
+        {
+            path: "/home",
+            redirect: "/"
+        },
+        {
+            path: "/quiz/:id",
+            name: "quiz",
+            component: QuizView
+        },
+        {
+            path: "/:catchall(.*)*",
+            name: "Not Found",
+            component: PageNotFoundView
+        }
+    ]
+})
+
+
+export default router;
