+++ QuizApp/src/views/404View.vue
@@ -1,0 +1,5 @@
+<template>
+    <div>
+        <h1>404 Page not found</h1>
+    </div>
+</template>
