+++ QuizApp/src/views/QuizView.vue
@@ -1,0 +1,5 @@
+<template>
+    <div>
+        <h1>QuizView</h1>
+    </div>
+</template>
